+++ onegov/election_day/forms/upload.py
@@ -3,7 +3,7 @@ from onegov.form import Form
 from onegov.form.fields import UploadField
 from onegov.form.parser.core import FieldDependency
 from onegov.form.validators import WhitelistedMimeType, FileSizeLimit
-from wtforms import RadioField
+from wtforms import BooleanField, RadioField
 from wtforms.validators import DataRequired, InputRequired
 from wtforms_components import If, Chain
 
@@ -15,10 +15,60 @@ ALLOWED_MIME_TYPES = {
     'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet'
 }
 
-MAX_FILE_SIZE = 5 * 1024 * 1024
+MAX_FILE_SIZE = 10 * 1024 * 1024
 
 
-class UploadForm(Form):
+class UploadElectionForm(Form):
+
+    type = RadioField(_("Type"), choices=[
+        ('sesam', _("SESAM")),
+        ('wabsti', _("Wabsti")),
+    ], validators=[InputRequired()], default='sesam')
+
+    # XXX make this easier with onegov.form
+    wabsti_dependency = FieldDependency('type', 'wabsti')
+    wabsti_validators = [If(
+        wabsti_dependency.fulfilled,
+        Chain((
+            WhitelistedMimeType(ALLOWED_MIME_TYPES),
+            FileSizeLimit(MAX_FILE_SIZE)
+        ))
+    )]
+
+    results = UploadField(
+        label=_("Results"),
+        validators=[
+            DataRequired(),
+            WhitelistedMimeType(ALLOWED_MIME_TYPES),
+            FileSizeLimit(MAX_FILE_SIZE)
+        ],
+        render_kw={'force_simple': True}
+    )
+
+    elected = UploadField(
+        label=_("Elected Candidates"),
+        validators=wabsti_validators,
+        render_kw=dict(force_simple=True, **wabsti_dependency.html_data)
+    )
+
+    statistics = UploadField(
+        label=_("Election statistics"),
+        validators=wabsti_validators,
+        render_kw=dict(force_simple=True, **wabsti_dependency.html_data)
+    )
+
+    complete = BooleanField(
+        render_kw=dict(force_simple=True, **wabsti_dependency.html_data)
+    )
+
+    def apply_model(self, model):
+        if model.type == 'majorz':
+            self.statistics.render_kw['data-depends-on'] = 'type/none'
+        else:
+            self.statistics.render_kw['data-depends-on'] = 'type/wabsti'
+
+
+class UploadVoteForm(Form):
 
     type = RadioField(_("Type"), choices=[
         ('simple', _("Simple Vote")),