+++ HISTORY.rst
@@ -1,6 +1,12 @@
 Changelog
 ---------
 
+- Adds elections.
+  [msom]
+
+- Adds access to all elections and votes of an election day.
+  [msom]
+
 0.6.0 (2016-02-16)
 ~~~~~~~~~~~~~~~~~~~